+++ climpred/tests/test_hindcastEnsemble_class.py
@@ -121,6 +121,23 @@ def test_inplace(
     assert hindcast != summed
 
 
+@pytest.mark.parametrize('alignment', ['same_inits', 'same_verifs', 'maximize'])
+def test_mean_reduce_bias(hindcast_hist_obs_1d, alignment):
+    how = 'mean'
+    metric = 'rmse'
+    hindcast = hindcast_hist_obs_1d
+    biased_skill = hindcast.verify(metric=metric, alignment=alignment)
+    bias_reduced_skill = hindcast.reduce_bias(
+        how=how, alignment=alignment, cross_validate=False
+    ).verify(metric=metric, alignment=alignment)
+    bias_reduced_skill_properly = hindcast.reduce_bias(
+        how=how, cross_validate=True, alignment=alignment
+    ).verify(metric=metric, alignment=alignment)
+    assert biased_skill > bias_reduced_skill
+    assert biased_skill > bias_reduced_skill_properly
+    assert bias_reduced_skill_properly >= bias_reduced_skill
+
+
 def test_verify_metric_kwargs(hindcast_hist_obs_1d):
     """Test that HindcastEnsemble works with metrics using metric_kwargs."""
     assert hindcast_hist_obs_1d.verify(